LayoutTests/inspector/model/remote-object/iterator-expected.txt


-----------------------------------------------------
EXPRESSION: "a"[Symbol.iterator]()
{
  "_type": "object",
  "_subtype": "iterator",
  "_objectId": "<filtered>",
  "_description": "String Iterator",
  "_preview": {
    "_type": "object",
    "_subtype": "iterator",
    "_description": "String Iterator",
    "_lossless": true,
    "_overflow": false,
    "_properties": [
      {
        "_name": "string",
        "_type": "string",
        "_value": "a",
        "_internal": true
      }
    ],
    "_entries": [
      {
        "_value": {
          "_type": "string",
          "_description": "a",
          "_lossless": true,
          "_overflow": false,
          "_properties": null,
          "_entries": null
        }
      }
    ]
  }
}

-----------------------------------------------------
EXPRESSION: "long string"[Symbol.iterator]()
{
  "_type": "object",
  "_subtype": "iterator",
  "_objectId": "<filtered>",
  "_description": "String Iterator",
  "_preview": {
    "_type": "object",
    "_subtype": "iterator",
    "_description": "String Iterator",
    "_lossless": false,
    "_overflow": true,
    "_properties": [
      {
        "_name": "string",
        "_type": "string",
        "_value": "long string",
        "_internal": true
      }
    ],
    "_entries": [
      {
        "_value": {
          "_type": "string",
          "_description": "l",
          "_lossless": true,
          "_overflow": false,
          "_properties": null,
          "_entries": null
        }
      },
      {
        "_value": {
          "_type": "string",
          "_description": "o",
          "_lossless": true,
          "_overflow": false,
          "_properties": null,
          "_entries": null
        }
      },
      {
        "_value": {
          "_type": "string",
          "_description": "n",
          "_lossless": true,
          "_overflow": false,
          "_properties": null,
          "_entries": null
        }
      },
      {
        "_value": {
          "_type": "string",
          "_description": "g",
          "_lossless": true,
          "_overflow": false,
          "_properties": null,
          "_entries": null
        }
      },
      {
        "_value": {
          "_type": "string",
          "_description": " ",
          "_lossless": true,
          "_overflow": false,
          "_properties": null,
          "_entries": null
        }
      }
    ]
  }
}

-----------------------------------------------------
EXPRESSION: [][Symbol.iterator]()
{
  "_type": "object",
  "_subtype": "iterator",
  "_objectId": "<filtered>",
  "_description": "Array Iterator",
  "_preview": {
    "_type": "object",
    "_subtype": "iterator",
    "_description": "Array Iterator",
    "_lossless": true,
    "_overflow": false,
    "_properties": [
      {
        "_name": "array",
        "_type": "object",
        "_subtype": "array",
        "_valuePreview": {
          "_type": "object",
          "_subtype": "array",
          "_description": "Array",
          "_lossless": true,
          "_overflow": false,
          "_size": 0,
          "_properties": [],
          "_entries": null
        },
        "_internal": true
      },
      {
        "_name": "kind",
        "_type": "string",
        "_value": "values",
        "_internal": true
      }
    ],
    "_entries": []
  }
}

-----------------------------------------------------
EXPRESSION: [1][Symbol.iterator]()
{
  "_type": "object",
  "_subtype": "iterator",
  "_objectId": "<filtered>",
  "_description": "Array Iterator",
  "_preview": {
    "_type": "object",
    "_subtype": "iterator",
    "_description": "Array Iterator",
    "_lossless": true,
    "_overflow": false,
    "_properties": [
      {
        "_name": "array",
        "_type": "object",
        "_subtype": "array",
        "_valuePreview": {
          "_type": "object",
          "_subtype": "array",
          "_description": "Array",
          "_lossless": true,
          "_overflow": false,
          "_size": 1,
          "_properties": [
            {
              "_name": "0",
              "_type": "number",
              "_value": "1"
            }
          ],
          "_entries": null
        },
        "_internal": true
      },
      {
        "_name": "kind",
        "_type": "string",
        "_value": "values",
        "_internal": true
      }
    ],
    "_entries": [
      {
        "_value": {
          "_type": "number",
          "_description": "1",
          "_lossless": true,
          "_overflow": false,
          "_properties": null,
          "_entries": null
        }
      }
    ]
  }
}

-----------------------------------------------------
EXPRESSION: [1, "two", 3, "four", 5, "size"][Symbol.iterator]()
{
  "_type": "object",
  "_subtype": "iterator",
  "_objectId": "<filtered>",
  "_description": "Array Iterator",
  "_preview": {
    "_type": "object",
    "_subtype": "iterator",
    "_description": "Array Iterator",
    "_lossless": false,
    "_overflow": true,
    "_properties": [
      {
        "_name": "array",
        "_type": "object",
        "_subtype": "array",
        "_value": "Array",
        "_internal": true
      },
      {
        "_name": "kind",
        "_type": "string",
        "_value": "values",
        "_internal": true
      }
    ],
    "_entries": [
      {
        "_value": {
          "_type": "number",
          "_description": "1",
          "_lossless": true,
          "_overflow": false,
          "_properties": null,
          "_entries": null
        }
      },
      {
        "_value": {
          "_type": "string",
          "_description": "two",
          "_lossless": true,
          "_overflow": false,
          "_properties": null,
          "_entries": null
        }
      },
      {
        "_value": {
          "_type": "number",
          "_description": "3",
          "_lossless": true,
          "_overflow": false,
          "_properties": null,
          "_entries": null
        }
      },
      {
        "_value": {
          "_type": "string",
          "_description": "four",
          "_lossless": true,
          "_overflow": false,
          "_properties": null,
          "_entries": null
        }
      },
      {
        "_value": {
          "_type": "number",
          "_description": "5",
          "_lossless": true,
          "_overflow": false,
          "_properties": null,
          "_entries": null
        }
      }
    ]
  }
}

-----------------------------------------------------
EXPRESSION: [1, "two", 3, "four", 5, "size"].keys()
{
  "_type": "object",
  "_subtype": "iterator",
  "_objectId": "<filtered>",
  "_description": "Array Iterator",
  "_preview": {
    "_type": "object",
    "_subtype": "iterator",
    "_description": "Array Iterator",
    "_lossless": false,
    "_overflow": true,
    "_properties": [
      {
        "_name": "array",
        "_type": "object",
        "_subtype": "array",
        "_value": "Array",
        "_internal": true
      },
      {
        "_name": "kind",
        "_type": "string",
        "_value": "keys",
        "_internal": true
      }
    ],
    "_entries": [
      {
        "_value": {
          "_type": "number",
          "_description": "0",
          "_lossless": true,
          "_overflow": false,
          "_properties": null,
          "_entries": null
        }
      },
      {
        "_value": {
          "_type": "number",
          "_description": "1",
          "_lossless": true,
          "_overflow": false,
          "_properties": null,
          "_entries": null
        }
      },
      {
        "_value": {
          "_type": "number",
          "_description": "2",
          "_lossless": true,
          "_overflow": false,
          "_properties": null,
          "_entries": null
        }
      },
      {
        "_value": {
          "_type": "number",
          "_description": "3",
          "_lossless": true,
          "_overflow": false,
          "_properties": null,
          "_entries": null
        }
      },
      {
        "_value": {
          "_type": "number",
          "_description": "4",
          "_lossless": true,
          "_overflow": false,
          "_properties": null,
          "_entries": null
        }
      }
    ]
  }
}

-----------------------------------------------------
EXPRESSION: [1, "two", 3, "four", 5, "size"].entries()
{
  "_type": "object",
  "_subtype": "iterator",
  "_objectId": "<filtered>",
  "_description": "Array Iterator",
  "_preview": {
    "_type": "object",
    "_subtype": "iterator",
    "_description": "Array Iterator",
    "_lossless": false,
    "_overflow": true,
    "_properties": [
      {
        "_name": "array",
        "_type": "object",
        "_subtype": "array",
        "_value": "Array",
        "_internal": true
      },
      {
        "_name": "kind",
        "_type": "string",
        "_value": "entries",
        "_internal": true
      }
    ],
    "_entries": [
      {
        "_value": {
          "_type": "object",
          "_subtype": "array",
          "_description": "Array",
          "_lossless": true,
          "_overflow": false,
          "_size": 2,
          "_properties": [
            {
              "_name": "0",
              "_type": "number",
              "_value": "0"
            },
            {
              "_name": "1",
              "_type": "number",
              "_value": "1"
            }
          ],
          "_entries": null
        }
      },
      {
        "_value": {
          "_type": "object",
          "_subtype": "array",
          "_description": "Array",
          "_lossless": true,
          "_overflow": false,
          "_size": 2,
          "_properties": [
            {
              "_name": "0",
              "_type": "number",
              "_value": "1"
            },
            {
              "_name": "1",
              "_type": "string",
              "_value": "two"
            }
          ],
          "_entries": null
        }
      },
      {
        "_value": {
          "_type": "object",
          "_subtype": "array",
          "_description": "Array",
          "_lossless": true,
          "_overflow": false,
          "_size": 2,
          "_properties": [
            {
              "_name": "0",
              "_type": "number",
              "_value": "2"
            },
            {
              "_name": "1",
              "_type": "number",
              "_value": "3"
            }
          ],
          "_entries": null
        }
      },
      {
        "_value": {
          "_type": "object",
          "_subtype": "array",
          "_description": "Array",
          "_lossless": true,
          "_overflow": false,
          "_size": 2,
          "_properties": [
            {
              "_name": "0",
              "_type": "number",
              "_value": "3"
            },
            {
              "_name": "1",
              "_type": "string",
              "_value": "four"
            }
          ],
          "_entries": null
        }
      },
      {
        "_value": {
          "_type": "object",
          "_subtype": "array",
          "_description": "Array",
          "_lossless": true,
          "_overflow": false,
          "_size": 2,
          "_properties": [
            {
              "_name": "0",
              "_type": "number",
              "_value": "4"
            },
            {
              "_name": "1",
              "_type": "number",
              "_value": "5"
            }
          ],
          "_entries": null
        }
      }
    ]
  }
}

-----------------------------------------------------
EXPRESSION: map = new Map; map.set(1, 2); map.set("key", "value"); map.values()
{
  "_type": "object",
  "_subtype": "iterator",
  "_objectId": "<filtered>",
  "_description": "Map Iterator",
  "_preview": {
    "_type": "object",
    "_subtype": "iterator",
    "_description": "Map Iterator",
    "_lossless": false,
    "_overflow": false,
    "_properties": [
      {
        "_name": "map",
        "_type": "object",
        "_subtype": "map",
        "_value": "Map",
        "_internal": true
      },
      {
        "_name": "kind",
        "_type": "string",
        "_value": "values",
        "_internal": true
      }
    ],
    "_entries": [
      {
        "_value": {
          "_type": "number",
          "_description": "2",
          "_lossless": true,
          "_overflow": false,
          "_properties": null,
          "_entries": null
        }
      },
      {
        "_value": {
          "_type": "string",
          "_description": "value",
          "_lossless": true,
          "_overflow": false,
          "_properties": null,
          "_entries": null
        }
      }
    ]
  }
}

-----------------------------------------------------
EXPRESSION: map.keys()
{
  "_type": "object",
  "_subtype": "iterator",
  "_objectId": "<filtered>",
  "_description": "Map Iterator",
  "_preview": {
    "_type": "object",
    "_subtype": "iterator",
    "_description": "Map Iterator",
    "_lossless": false,
    "_overflow": false,
    "_properties": [
      {
        "_name": "map",
        "_type": "object",
        "_subtype": "map",
        "_value": "Map",
        "_internal": true
      },
      {
        "_name": "kind",
        "_type": "string",
        "_value": "keys",
        "_internal": true
      }
    ],
    "_entries": [
      {
        "_value": {
          "_type": "number",
          "_description": "1",
          "_lossless": true,
          "_overflow": false,
          "_properties": null,
          "_entries": null
        }
      },
      {
        "_value": {
          "_type": "string",
          "_description": "key",
          "_lossless": true,
          "_overflow": false,
          "_properties": null,
          "_entries": null
        }
      }
    ]
  }
}

-----------------------------------------------------
EXPRESSION: map.entries()
{
  "_type": "object",
  "_subtype": "iterator",
  "_objectId": "<filtered>",
  "_description": "Map Iterator",
  "_preview": {
    "_type": "object",
    "_subtype": "iterator",
    "_description": "Map Iterator",
    "_lossless": false,
    "_overflow": false,
    "_properties": [
      {
        "_name": "map",
        "_type": "object",
        "_subtype": "map",
        "_value": "Map",
        "_internal": true
      },
      {
        "_name": "kind",
        "_type": "string",
        "_value": "entries",
        "_internal": true
      }
    ],
    "_entries": [
      {
        "_value": {
          "_type": "object",
          "_subtype": "array",
          "_description": "Array",
          "_lossless": true,
          "_overflow": false,
          "_size": 2,
          "_properties": [
            {
              "_name": "0",
              "_type": "number",
              "_value": "1"
            },
            {
              "_name": "1",
              "_type": "number",
              "_value": "2"
            }
          ],
          "_entries": null
        }
      },
      {
        "_value": {
          "_type": "object",
          "_subtype": "array",
          "_description": "Array",
          "_lossless": true,
          "_overflow": false,
          "_size": 2,
          "_properties": [
            {
              "_name": "0",
              "_type": "string",
              "_value": "key"
            },
            {
              "_name": "1",
              "_type": "string",
              "_value": "value"
            }
          ],
          "_entries": null
        }
      }
    ]
  }
}

-----------------------------------------------------
EXPRESSION: set = new Set; for (var i = 0; i <= 100; i++) set.add(i); set.values()
{
  "_type": "object",
  "_subtype": "iterator",
  "_objectId": "<filtered>",
  "_description": "Set Iterator",
  "_preview": {
    "_type": "object",
    "_subtype": "iterator",
    "_description": "Set Iterator",
    "_lossless": false,
    "_overflow": true,
    "_properties": [
      {
        "_name": "set",
        "_type": "object",
        "_subtype": "set",
        "_value": "Set",
        "_internal": true
      },
      {
        "_name": "kind",
        "_type": "string",
        "_value": "values",
        "_internal": true
      }
    ],
    "_entries": [
      {
        "_value": {
          "_type": "number",
          "_description": "0",
          "_lossless": true,
          "_overflow": false,
          "_properties": null,
          "_entries": null
        }
      },
      {
        "_value": {
          "_type": "number",
          "_description": "1",
          "_lossless": true,
          "_overflow": false,
          "_properties": null,
          "_entries": null
        }
      },
      {
        "_value": {
          "_type": "number",
          "_description": "2",
          "_lossless": true,
          "_overflow": false,
          "_properties": null,
          "_entries": null
        }
      },
      {
        "_value": {
          "_type": "number",
          "_description": "3",
          "_lossless": true,
          "_overflow": false,
          "_properties": null,
          "_entries": null
        }
      },
      {
        "_value": {
          "_type": "number",
          "_description": "4",
          "_lossless": true,
          "_overflow": false,
          "_properties": null,
          "_entries": null
        }
      }
    ]
  }
}

-----------------------------------------------------
EXPRESSION: map.entries()
{
  "_type": "object",
  "_subtype": "iterator",
  "_objectId": "<filtered>",
  "_description": "Map Iterator",
  "_preview": {
    "_type": "object",
    "_subtype": "iterator",
    "_description": "Map Iterator",
    "_lossless": false,
    "_overflow": false,
    "_properties": [
      {
        "_name": "map",
        "_type": "object",
        "_subtype": "map",
        "_value": "Map",
        "_internal": true
      },
      {
        "_name": "kind",
        "_type": "string",
        "_value": "entries",
        "_internal": true
      }
    ],
    "_entries": [
      {
        "_value": {
          "_type": "object",
          "_subtype": "array",
          "_description": "Array",
          "_lossless": true,
          "_overflow": false,
          "_size": 2,
          "_properties": [
            {
              "_name": "0",
              "_type": "number",
              "_value": "1"
            },
            {
              "_name": "1",
              "_type": "number",
              "_value": "2"
            }
          ],
          "_entries": null
        }
      },
      {
        "_value": {
          "_type": "object",
          "_subtype": "array",
          "_description": "Array",
          "_lossless": true,
          "_overflow": false,
          "_size": 2,
          "_properties": [
            {
              "_name": "0",
              "_type": "string",
              "_value": "key"
            },
            {
              "_name": "1",
              "_type": "string",
              "_value": "value"
            }
          ],
          "_entries": null
        }
      }
    ]
  }
}

-----------------------------------------------------
EXPRESSION: x = undefined; (function() { x = arguments; })(1, "two"); x[Symbol.iterator]()
{
  "_type": "object",
  "_subtype": "iterator",
  "_objectId": "<filtered>",
  "_description": "Array Iterator",
  "_preview": {
    "_type": "object",
    "_subtype": "iterator",
    "_description": "Array Iterator",
    "_lossless": true,
    "_overflow": false,
    "_properties": [
      {
        "_name": "array",
        "_type": "object",
        "_subtype": "array",
        "_valuePreview": {
          "_type": "object",
          "_subtype": "array",
          "_description": "Arguments",
          "_lossless": true,
          "_overflow": false,
          "_size": 2,
          "_properties": [
            {
              "_name": "0",
              "_type": "number",
              "_value": "1"
            },
            {
              "_name": "1",
              "_type": "string",
              "_value": "two"
            }
          ],
          "_entries": null
        },
        "_internal": true
      },
      {
        "_name": "kind",
        "_type": "string",
        "_value": "values",
        "_internal": true
      }
    ],
    "_entries": [
      {
        "_value": {
          "_type": "number",
          "_description": "1",
          "_lossless": true,
          "_overflow": false,
          "_properties": null,
          "_entries": null
        }
      },
      {
        "_value": {
          "_type": "string",
          "_description": "two",
          "_lossless": true,
          "_overflow": false,
          "_properties": null,
          "_entries": null
        }
      }
    ]
  }
}

-----------------------------------------------------
EXPRESSION: iter = [1, 2][Symbol.iterator](); iter["return"] = function(){}; iter
{
  "_type": "object",
  "_subtype": "iterator",
  "_objectId": "<filtered>",
  "_description": "Array Iterator",
  "_preview": {
    "_type": "object",
    "_subtype": "iterator",
    "_description": "Array Iterator",
    "_lossless": true,
    "_overflow": false,
    "_properties": [
      {
        "_name": "array",
        "_type": "object",
        "_subtype": "array",
        "_valuePreview": {
          "_type": "object",
          "_subtype": "array",
          "_description": "Array",
          "_lossless": true,
          "_overflow": false,
          "_size": 2,
          "_properties": [
            {
              "_name": "0",
              "_type": "number",
              "_value": "1"
            },
            {
              "_name": "1",
              "_type": "number",
              "_value": "2"
            }
          ],
          "_entries": null
        },
        "_internal": true
      },
      {
        "_name": "kind",
        "_type": "string",
        "_value": "values",
        "_internal": true
      }
    ],
    "_entries": [
      {
        "_value": {
          "_type": "number",
          "_description": "1",
          "_lossless": true,
          "_overflow": false,
          "_properties": null,
          "_entries": null
        }
      },
      {
        "_value": {
          "_type": "number",
          "_description": "2",
          "_lossless": true,
          "_overflow": false,
          "_properties": null,
          "_entries": null
        }
      }
    ]
  }
}
